認証機能を追加

+++ app/controllers/api/v1/links_controller.rb
@@ -1,30 +1,41 @@
 class Api::V1::LinksController < ApplicationController
+  before_action :authenticate_api_v1_user!, only: %i[create update destroy]
+
   def index
     links = Link.all
-    render json: links.as_json
+    render status: :ok, json: { status: 200, data: links }
   end
 
   def create
-    link = Link.new(link_params)
-    if link.save
-      render json: link.as_json
+    if api_v1_user_signed_in?
+      link = current_api_v1_user.links.build(link_params)
+      if link.save
+        render status: :ok, json: { status: 200, data: link }
+      else
+        render status: :bad_request, json: { status: 400, data: link.errors }
+      end
     else
-      render json: { data: link.errors }
+      render json: { message: "ログインしてください" }
     end
   end
 
   def update
     link = Link.find(params[:id])
     if link.update(link_params)
-      render json: link.as_json
+      render status: :ok, json: { status: 200, data: link }
     else
-      render json: { data: link.errors }
+      render status: :not_found, json: { status: 404, data: link.errors, message: "更新に失敗しました" }
     end
   end
 
   def destroy
     link = Link.find(params[:id])
-    link.destroy
+    if current_api_v1_user.id == link.user_id
+      link.destroy
+      render status: :ok, json: { status: 200, message: "削除に成功しました" }
+    else
+      render status: :not_found, json: { status: 404, message: "削除に失敗しました" }
+    end
   end
 
   private
